+++ pkg/linter/lib/src/rules/unnecessary_type_name_in_constructor.dart
@@ -1,0 +1,55 @@
+// Copyright (c) 2026, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:analyzer/analysis_rule/analysis_rule.dart';
+import 'package:analyzer/analysis_rule/rule_context.dart';
+import 'package:analyzer/analysis_rule/rule_visitor_registry.dart';
+import 'package:analyzer/dart/analysis/features.dart';
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/visitor.dart';
+import 'package:analyzer/error/error.dart';
+
+import '../analyzer.dart';
+import '../diagnostic.dart' as diag;
+
+const _desc = "Don't use an explicit type name in a constructor.";
+
+class UnnecessaryTypeNameInConstructor extends AnalysisRule {
+  UnnecessaryTypeNameInConstructor()
+    : super(
+        name: LintNames.unnecessary_type_name_in_constructor,
+        description: _desc,
+      );
+
+  @override
+  bool get canUseParsedResult => true;
+
+  @override
+  DiagnosticCode get diagnosticCode => diag.unnecessaryTypeNameInConstructor;
+
+  @override
+  void registerNodeProcessors(
+    RuleVisitorRegistry registry,
+    RuleContext context,
+  ) {
+    if (!context.isFeatureEnabled(Feature.primary_constructors)) {
+      return;
+    }
+    var visitor = _Visitor(this);
+    registry.addConstructorDeclaration(this, visitor);
+  }
+}
+
+class _Visitor extends SimpleAstVisitor<void> {
+  final AnalysisRule rule;
+  _Visitor(this.rule);
+
+  @override
+  void visitConstructorDeclaration(ConstructorDeclaration node) {
+    var typeName = node.typeName;
+    if (typeName != null) {
+      rule.reportAtNode(typeName);
+    }
+  }
+}