+++ pkg/linter/test/rules/use_late_for_private_fields_and_variables_test.dart
@@ -54,21 +54,36 @@ class C {
     );
   }
 
+  test_instanceField_private_assignedInConstructorInitializer() async {
+    await assertNoDiagnostics('''
+class C {
+  C([int? i]) : _i = i;
+  int? _i;
+}
+''');
+  }
+
+  test_instanceField_private_assignedInPrimaryConstructorInitializer() async {
+    await assertNoDiagnostics('''
+class C([int? i]) {
+  this : _i = i;
+  int? _i;
+}
+''');
+  }
+
   test_instanceField_private_declaredInPart() async {
     newFile('$testPackageLibPath/lib.dart', r'''
 part 'test.dart';
 ''');
-    await assertDiagnostics(
-      '''
+    await assertNoDiagnostics('''
 part of 'lib.dart';
 
 class C {
   final String? _s;
   C(this._s);
 }
-''',
-      [lint(47, 2)],
-    );
+''');
   }
 
   /// https://github.com/dart-lang/linter/issues/3823
@@ -120,6 +135,37 @@ class C {
     await assertNoDiagnostics('''
 class const C([this._i]) {
   final int? _i;
+}
+''');
+  }
+
+  test_instanceField_private_primaryConstructor() async {
+    await assertNoDiagnostics('''
+class C(int? _i);
+''');
+  }
+
+  test_instanceField_private_primaryConstructor_namedParameter() async {
+    await assertNoDiagnostics('''
+// ignore: unused_field_from_primary_constructor
+class C({final int? _i});
+''');
+  }
+
+  test_instanceField_private_withFieldFormalParameter() async {
+    await assertNoDiagnostics('''
+class C {
+  C([this._i]);
+  int? _i;
+}
+''');
+  }
+
+  test_instanceField_private_withFieldFormalParameter_named() async {
+    await assertNoDiagnostics('''
+class C {
+  C({this._i});
+  int? _i;
 }
 ''');
   }